+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-## Unreleased
+## [v1.6.0]
+
+**Release date: 2021-09-30**
 
 ### Added
 
@@ -1344,6 +1346,7 @@ There was no CHANGELOG file prior to this release, so I don't have much
 information about changes, except for
 [commit messages](../../commits/v0.2).
 
+[v1.6.0]: ../../releases/tag/v1.6.0
 [v1.5.1]: ../../releases/tag/v1.5.1
 [v1.5.0]: ../../releases/tag/v1.5.0
 [v1.4.1]: ../../releases/tag/v1.4.1